Do not include the CLI parsing code into the shared object.
